Increase tickrate and send interval of network
Add some more smoothness to the movement and switch from simple
transform teleporting to character controller movement, enabling
collisions. Add a compressed input structure to optimze the data sent in
commands. Add a tool to build the game quickly and also quickstart the
game to connect immediately to the server